Refactored Goldilocks
Moved the Goldilocks specific code into the Goldilocks module.
