+++ goldilocks/Cargo.toml
@@ -7,11 +7,15 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 p3-field = { path = "../field" }
 p3-util = { path = "../util" }
+p3-poseidon2 = { path = "../poseidon2" }
+p3-symmetric = { path = "../symmetric" }
 rand = "0.8.5"
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 
 [dev-dependencies]
 p3-field-testing = { path = "../field-testing" }
+ark-ff = { version = "^0.4.0", default-features = false }
+zkhash = { git = "https://github.com/HorizenLabs/poseidon2" }
 criterion = "0.5.1"
 
 [[bench]]
