+++ poseidon2/Cargo.toml
@@ -5,16 +5,14 @@ edition = "2021"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
-p3-goldilocks = { path = "../goldilocks" }
-p3-mersenne-31 = { path = "../mersenne-31" }
 p3-field = { path = "../field" }
 p3-symmetric = { path = "../symmetric" }
 rand = { version = "0.8.5", features = ["min_const_gen"] }
 
 [dev-dependencies]
+p3-mersenne-31 = { path = "../mersenne-31" }
 p3-baby-bear = { path = "../baby-bear" }
-ark-ff = { version = "^0.4.0", default-features = false }
-zkhash = { git = "https://github.com/HorizenLabs/poseidon2" }
+p3-goldilocks = { path = "../goldilocks" }
 criterion = "0.5.1"
 
 [[bench]]
