+++ poseidon2/benches/poseidon2.rs
@@ -3,9 +3,9 @@ use std::any::type_name;
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use p3_baby_bear::{BabyBear, DiffusionMatrixBabybear};
 use p3_field::PrimeField64;
-use p3_goldilocks::Goldilocks;
+use p3_goldilocks::{Goldilocks, DiffusionMatrixGoldilocks};
 use p3_poseidon2::{
-    DiffusionMatrixGoldilocks, DiffusionPermutation, Poseidon2,
+    DiffusionPermutation, Poseidon2,
 };
 use p3_symmetric::Permutation;
 use rand::distributions::{Distribution, Standard};
