+++ poseidon2/src/lib.rs
@@ -9,12 +9,10 @@
 extern crate alloc;
 
 mod diffusion;
-mod goldilocks;
 mod matrix;
 use alloc::vec::Vec;
 
-pub use diffusion::DiffusionPermutation;
-pub use goldilocks::DiffusionMatrixGoldilocks;
+pub use diffusion::{DiffusionPermutation, matmul_internal};
 
 use matrix::Poseidon2MEMatrix;
 use p3_field::{AbstractField, PrimeField};
@@ -158,165 +156,4 @@ where
     AF::F: PrimeField,
     Diffusion: DiffusionPermutation<AF, WIDTH>,
 {
-}
-
-#[cfg(test)]
-mod tests {
-    use alloc::vec::Vec;
-
-    use ark_ff::{BigInteger, PrimeField};
-    use p3_field::AbstractField;
-    use p3_goldilocks::Goldilocks;
-    use p3_symmetric::Permutation;
-    use rand::Rng;
-    use zkhash::fields::goldilocks::FpGoldiLocks;
-    use zkhash::poseidon2::poseidon2::Poseidon2 as Poseidon2Ref;  
-    use zkhash::poseidon2::poseidon2_instance_goldilocks::{
-        POSEIDON2_GOLDILOCKS_12_PARAMS, POSEIDON2_GOLDILOCKS_8_PARAMS, RC12, RC8,
-    };
-
-    use crate::goldilocks::DiffusionMatrixGoldilocks;
-    use crate::{Poseidon2};
-
-    fn goldilocks_from_ark_ff(input: FpGoldiLocks) -> Goldilocks {
-        let as_bigint = input.into_bigint();
-        let mut as_bytes = as_bigint.to_bytes_le();
-        as_bytes.resize(8, 0);
-        let as_u64 = u64::from_le_bytes(as_bytes[0..8].try_into().unwrap());
-        Goldilocks::from_wrapped_u64(as_u64)
-    }
-
-    #[test]
-    fn test_poseidon2_goldilocks_width_8() {
-        const WIDTH: usize = 8;
-        const D: u64 = 7;
-        const ROUNDS_F: usize = 8;
-        const ROUNDS_P: usize = 22;
-
-        type F = Goldilocks;
-
-        let mut rng = rand::thread_rng();
-
-        // Poiseidon2 reference implementation from zkhash repo.
-        let poseidon2_ref = Poseidon2Ref::new(&POSEIDON2_GOLDILOCKS_8_PARAMS);
-
-        // Copy over round constants from zkhash.
-        let round_constants: Vec<[F; WIDTH]> = RC8
-            .iter()
-            .map(|vec| {
-                vec.iter()
-                    .cloned()
-                    .map(goldilocks_from_ark_ff)
-                    .collect::<Vec<_>>()
-                    .try_into()
-                    .unwrap()
-            })
-            .collect();
-
-        // Our Poseidon2 implementation.
-        let poseidon2: Poseidon2<Goldilocks, DiffusionMatrixGoldilocks, WIDTH, D> = Poseidon2::new(
-            ROUNDS_F,
-            ROUNDS_P,
-            round_constants,
-            DiffusionMatrixGoldilocks,
-        );
-
-        // Generate random input and convert to both Goldilocks field formats.
-        let input_u64 = rng.gen::<[u64; WIDTH]>();
-        let input_ref = input_u64
-            .iter()
-            .cloned()
-            .map(FpGoldiLocks::from)
-            .collect::<Vec<_>>();
-        let input = input_u64.map(F::from_wrapped_u64);
-
-        // Check that the conversion is correct.
-        assert!(input_ref
-            .iter()
-            .zip(input.iter())
-            .all(|(a, b)| goldilocks_from_ark_ff(*a) == *b));
-
-        // Run reference implementation.
-        let output_ref = poseidon2_ref.permutation(&input_ref);
-        let expected: [F; WIDTH] = output_ref
-            .iter()
-            .cloned()
-            .map(goldilocks_from_ark_ff)
-            .collect::<Vec<_>>()
-            .try_into()
-            .unwrap();
-
-        // Run our implementation.
-        let mut output = input;
-        poseidon2.permute_mut(&mut output);
-
-        assert_eq!(output, expected);
-    }
-
-    #[test]
-    fn test_poseidon2_goldilocks_width_12() {
-        const WIDTH: usize = 12;
-        const D: u64 = 7;
-        const ROUNDS_F: usize = 8;
-        const ROUNDS_P: usize = 22;
-
-        type F = Goldilocks;
-
-        let mut rng = rand::thread_rng();
-
-        // Poiseidon2 reference implementation from zkhash repo.
-        let poseidon2_ref = Poseidon2Ref::new(&POSEIDON2_GOLDILOCKS_12_PARAMS);
-
-        // Copy over round constants from zkhash.
-        let round_constants: Vec<[F; WIDTH]> = RC12
-            .iter()
-            .map(|vec| {
-                vec.iter()
-                    .cloned()
-                    .map(goldilocks_from_ark_ff)
-                    .collect::<Vec<_>>()
-                    .try_into()
-                    .unwrap()
-            })
-            .collect();
-
-        // Our Poseidon2 implementation.
-        let poseidon2: Poseidon2<Goldilocks, DiffusionMatrixGoldilocks, WIDTH, D> = Poseidon2::new(
-            ROUNDS_F,
-            ROUNDS_P,
-            round_constants,
-            DiffusionMatrixGoldilocks,
-        );
-
-        // Generate random input and convert to both Goldilocks field formats.
-        let input_u64 = rng.gen::<[u64; WIDTH]>();
-        let input_ref = input_u64
-            .iter()
-            .cloned()
-            .map(FpGoldiLocks::from)
-            .collect::<Vec<_>>();
-        let input = input_u64.map(F::from_wrapped_u64);
-
-        // Check that the conversion is correct.
-        assert!(input_ref
-            .iter()
-            .zip(input.iter())
-            .all(|(a, b)| goldilocks_from_ark_ff(*a) == *b));
-
-        // Run reference implementation.
-        let output_ref = poseidon2_ref.permutation(&input_ref);
-        let expected: [F; WIDTH] = output_ref
-            .iter()
-            .cloned()
-            .map(goldilocks_from_ark_ff)
-            .collect::<Vec<_>>()
-            .try_into()
-            .unwrap();
-
-        // Run our implementation.
-        let mut output = input;
-        poseidon2.permute_mut(&mut output);
-
-        assert_eq!(output, expected);
-    }
 }
